Implement basic regex parser
The core parser for the initial release is going to be one where the user
specifies a regular expression for the incoming notification message. Obviously
this is aimed at relatively technically sound users. I'll be thinking of
implementing something more user-friendly than a full regex implementation
for future releases beyond v0.1.

Parser simply takes in a regex pattern, and matches the entire thing against
incoming messages.